Add a test for Admin Webmnetions Feed
For #452
